requirements.txt
aiogram>=3.0.0b3
aioredis~=2.0
environs~=9.0
Django
psycopg2-binary==2.8.6
psycopg2
pillow
python-dotenv
django-adminlte3
python-dateutil~=2.8.2